+++ modules/canvaskit/canvaskit_bindings.cpp
@@ -22,7 +22,6 @@
 #include "SkGradientShader.h"
 #include "SkImage.h"
 #include "SkImageInfo.h"
-#include "SkImageShader.h"
 #include "SkMakeUnique.h"
 #include "SkMaskFilter.h"
 #include "SkPaint.h"
@@ -632,22 +631,9 @@ EMSCRIPTEN_BINDINGS(Skia) {
 
         return SkImage::MakeRasterData(info, pixelData, rowBytes);
     }), allow_raw_pointers());
-    // Allow localMatrix to be optional, so we have 2 declarations of these shaders
-    function("_MakeImageShader", optional_override([](sk_sp<SkImage> img,
-                                SkTileMode tx, SkTileMode ty,
-                                bool clampAsIfUnpremul)->sk_sp<SkShader> {
-        return SkImageShader::Make(img, (SkTileMode)tx, (SkTileMode)ty, nullptr, clampAsIfUnpremul);
-    }), allow_raw_pointers());
-    function("_MakeImageShader", optional_override([](sk_sp<SkImage> img,
-                                SkTileMode tx, SkTileMode ty,
-                                bool clampAsIfUnpremul, const SimpleMatrix& lm)->sk_sp<SkShader> {
-        SkMatrix localMatrix = toSkMatrix(lm);
-
-        return SkImageShader::Make(img, (SkTileMode)tx, (SkTileMode)ty, &localMatrix, clampAsIfUnpremul);
-    }), allow_raw_pointers());
     function("_MakeLinearGradientShader", optional_override([](SkPoint start, SkPoint end,
                                 uintptr_t /* SkColor*  */ cPtr, uintptr_t /* SkScalar*  */ pPtr,
-                                int count, SkTileMode mode, uint32_t flags)->sk_sp<SkShader> {
+                                int count, SkShader::TileMode mode, uint32_t flags)->sk_sp<SkShader> {
         SkPoint points[] = { start, end };
         // See comment above for uintptr_t explanation
         const SkColor*  colors    = reinterpret_cast<const SkColor*> (cPtr);
@@ -658,7 +644,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
     }), allow_raw_pointers());
     function("_MakeLinearGradientShader", optional_override([](SkPoint start, SkPoint end,
                                 uintptr_t /* SkColor*  */ cPtr, uintptr_t /* SkScalar*  */ pPtr,
-                                int count, SkTileMode mode, uint32_t flags,
+                                int count, SkShader::TileMode mode, uint32_t flags,
                                 const SimpleMatrix& lm)->sk_sp<SkShader> {
         SkPoint points[] = { start, end };
         // See comment above for uintptr_t explanation
@@ -672,7 +658,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
     }), allow_raw_pointers());
     function("_MakeRadialGradientShader", optional_override([](SkPoint center, SkScalar radius,
                                 uintptr_t /* SkColor*  */ cPtr, uintptr_t /* SkScalar*  */ pPtr,
-                                int count, SkTileMode mode, uint32_t flags)->sk_sp<SkShader> {
+                                int count, SkShader::TileMode mode, uint32_t flags)->sk_sp<SkShader> {
         // See comment above for uintptr_t explanation
         const SkColor*  colors    = reinterpret_cast<const SkColor*> (cPtr);
         const SkScalar* positions = reinterpret_cast<const SkScalar*>(pPtr);
@@ -682,7 +668,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
     }), allow_raw_pointers());
     function("_MakeRadialGradientShader", optional_override([](SkPoint center, SkScalar radius,
                                 uintptr_t /* SkColor*  */ cPtr, uintptr_t /* SkScalar*  */ pPtr,
-                                int count, SkTileMode mode, uint32_t flags,
+                                int count, SkShader::TileMode mode, uint32_t flags,
                                 const SimpleMatrix& lm)->sk_sp<SkShader> {
         // See comment above for uintptr_t explanation
         const SkColor*  colors    = reinterpret_cast<const SkColor*> (cPtr);
@@ -696,7 +682,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
                 SkPoint start, SkScalar startRadius,
                 SkPoint end, SkScalar endRadius,
                 uintptr_t /* SkColor*  */ cPtr, uintptr_t /* SkScalar*  */ pPtr,
-                int count, SkTileMode mode, uint32_t flags)->sk_sp<SkShader> {
+                int count, SkShader::TileMode mode, uint32_t flags)->sk_sp<SkShader> {
         // See comment above for uintptr_t explanation
         const SkColor*  colors    = reinterpret_cast<const SkColor*> (cPtr);
         const SkScalar* positions = reinterpret_cast<const SkScalar*>(pPtr);
@@ -709,7 +695,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
                 SkPoint start, SkScalar startRadius,
                 SkPoint end, SkScalar endRadius,
                 uintptr_t /* SkColor*  */ cPtr, uintptr_t /* SkScalar*  */ pPtr,
-                int count, SkTileMode mode, uint32_t flags,
+                int count, SkShader::TileMode mode, uint32_t flags,
                 const SimpleMatrix& lm)->sk_sp<SkShader> {
         // See comment above for uintptr_t explanation
         const SkColor*  colors    = reinterpret_cast<const SkColor*> (cPtr);
@@ -920,7 +906,19 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .function("height", &SkImage::height)
         .function("width", &SkImage::width)
         .function("_encodeToData", select_overload<sk_sp<SkData>()const>(&SkImage::encodeToData))
-        .function("_encodeToDataWithFormat", select_overload<sk_sp<SkData>(SkEncodedImageFormat encodedImageFormat, int quality)const>(&SkImage::encodeToData));
+        .function("_encodeToDataWithFormat", select_overload<sk_sp<SkData>(SkEncodedImageFormat encodedImageFormat, int quality)const>(&SkImage::encodeToData))
+            // Allow localMatrix to be optional, so we have 2 declarations of these shaders
+        .function("_makeShader", optional_override([](sk_sp<SkImage> self,
+                                SkShader::TileMode tx, SkShader::TileMode ty)->sk_sp<SkShader> {
+            return self->makeShader((SkTileMode)tx, (SkTileMode)ty, nullptr);
+        }), allow_raw_pointers())
+        .function("_makeShader", optional_override([](sk_sp<SkImage> self,
+                                 SkShader::TileMode tx, SkShader::TileMode ty,
+                                 const SimpleMatrix& lm)->sk_sp<SkShader> {
+            SkMatrix localMatrix = toSkMatrix(lm);
+
+            return self->makeShader((SkTileMode)tx, (SkTileMode)ty, &localMatrix);
+        }), allow_raw_pointers());
 
     class_<SkMaskFilter>("SkMaskFilter")
         .smart_ptr<sk_sp<SkMaskFilter>>("sk_sp<SkMaskFilter>");
@@ -1193,12 +1191,12 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .value("UTF32",   SkTextEncoding::kUTF32)
         .value("GlyphID", SkTextEncoding::kGlyphID);
 
-    enum_<SkTileMode>("TileMode")
-        .value("Clamp",    SkTileMode::kClamp)
-        .value("Repeat",   SkTileMode::kRepeat)
-        .value("Mirror",   SkTileMode::kMirror)
+    enum_<SkShader::TileMode>("TileMode")
+        .value("Clamp",    SkShader::TileMode::kClamp_TileMode)
+        .value("Repeat",   SkShader::TileMode::kRepeat_TileMode)
+        .value("Mirror",   SkShader::TileMode::kMirror_TileMode)
         // Decal mode only works in the SW backend, not WebGl (yet).
-        .value("Decal",    SkTileMode::kDecal);
+        .value("Decal",    SkShader::TileMode::kDecal_TileMode);
 
     enum_<SkVertices::VertexMode>("VertexMode")
         .value("Triangles",       SkVertices::VertexMode::kTriangles_VertexMode)